pyproject.toml

[project]
name = "hmexpmatrix"
version = "0.1.0"
description = "Utility for setup of Harmonie experiment families"
authors = [
    {name = "Leif Denby", email = "lcd@dmi.dk"},
]
dependencies = [
    "jinja2>=3.1.2",
    "pyyaml>=6.0",
    "loguru>=0.7.0",
    "isodate>=0.6.1",
    "rich>=13.4.1",
]
requires-python = ">=3.9"
readme = "README.md"
license = {text = "MIT"}

[build-system]
requires = ["pdm-backend"]
build-backend = "pdm.backend"

[tool.pdm.dev-dependencies]
dev = [
    "pre-commit>=3.3.2",
]
